src/auth_service/requirements.txt
fastapi[all]
sqlalchemy
sqlalchemy-utils
psycopg2-binary
asyncpg
alembic
pydantic-extra-types
pydantic[email]
phonenumbers
passlib
pyjwt
redis